add 'sync' task
add a task that is used to create local copies of archived... archives. the main goal is to not lose them in case they get deleted, and to have an easy way of saving them. by default, the task uses `murolem/wplace-archives` repo (configurable; linked in first few lines of readme). each release is pulled, saved and validated. it's still wip, so it's pretty barebones. the task can only be run directly from the repo with `bun run sync` and NOT from a built binary. type `bun run sync --help` to see all the options.

- add 'sync; task.
- fix spawn still using stdout and stderr despite setting noInheritStdout option.
- split noInheritStdout into itself and noInheritStderr, since the former was handling both streams. less crap printing yippiee!!!
- add checks for command existence in all tasks for commands that are used in spawn's. should make them bail out early in case command (like 'gh') is missing.
- implement compound (multiple) prefixes for logger.
- implement cloning of logger instances.
- improve visibility of FATAL log level.
- implement zero-prefix support for logger.
- validate formatProgressToPercentage() inputs, handle 0 digits after comma.